move classes to component packages
