+++ app/src/main/AndroidManifest.xml
@@ -137,7 +137,7 @@
             android:name=".components.report.ReportActivity"
             android:windowSoftInputMode="stateAlwaysHidden|adjustResize" />
         <activity android:name=".components.instancemute.InstanceListActivity" />
-        <activity android:name=".ScheduledTootActivity" />
+        <activity android:name=".components.scheduled.ScheduledTootActivity" />
 
         <receiver android:name=".receiver.NotificationClearBroadcastReceiver" />
         <receiver
