+++ app/src/main/java/com/keylesspalace/tusky/components/scheduled/ScheduledTootAdapter.kt
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License along with Tusky; if not,
  * see <http://www.gnu.org/licenses>. */
 
-package com.keylesspalace.tusky.adapter
+package com.keylesspalace.tusky.components.scheduled
 
 import android.view.LayoutInflater
 import android.view.View
