+++ app/src/main/java/com/keylesspalace/tusky/di/ActivitiesModule.kt
@@ -19,6 +19,7 @@ import com.keylesspalace.tusky.*
 import com.keylesspalace.tusky.components.compose.ComposeActivity
 import com.keylesspalace.tusky.components.instancemute.InstanceListActivity
 import com.keylesspalace.tusky.components.report.ReportActivity
+import com.keylesspalace.tusky.components.scheduled.ScheduledTootActivity
 import com.keylesspalace.tusky.components.search.SearchActivity
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
